Adding SpawnPoints capable of spawning any number of enemies at any number of points
Added in EnemySpawnPoint.cs and an example of it being used.
Essentially you can just give it any enemy type that extends BasicEnemy
and it will instantiate it across the network at the positions passed
into it in the inspector. If no positions are passed in, it just spawns
them at it's own position.