Changing animations to be prompted by Input Axis and not Key.
See above. Small pause before switching to new animation due to some
funkiness with Unity, but we can iron that out in a bit. Punch is a bit
wonky, but we just need better overriding and it'll be alright. I'll
probably change spritesheets to have priorities and use a priority
queue for animating them.